// src/app/core/api/generated/schemas/adminReturnsControllerListRequestsParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export type AdminReturnsControllerListRequestsParams = {
  q?: string;
  status?: string;
  refund_status?: string;
  date_from?: string;
  date_to?: string;
  order_number?: string;
  return_number?: string;
  page?: number;
  limit?: number;
};