// src/app/core/api/generated/schemas/returnRequestListItemDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */
import type { ReturnRequestListItemDtoReasonName } from "./returnRequestListItemDtoReasonName";
import type { ReturnRequestListItemDtoSku } from "./returnRequestListItemDtoSku";

export interface ReturnRequestListItemDto {
  return_request_id: number;
  return_number: string;
  order_number: string;
  requested_at: string;
  status: string;
  refund_status: string;
  customer_name: string;
  product_title: string;
  /** @nullable */
  sku?: ReturnRequestListItemDtoSku;
  total_qty: number;
  /** @nullable */
  reason_name?: ReturnRequestListItemDtoReasonName;
}